requirements.txt
pycrypto
pyyaml
